requirements.txt
requests
atlassian-python-api
boto3
